feat: Implement Table component by handling query and API #33
